+++ spark_fhir_schemas/r4/complex_types/structuremap_source.py
@@ -5,6 +5,7 @@ from typing import Union
 from pyspark.sql.types import ArrayType
 from pyspark.sql.types import BooleanType
 from pyspark.sql.types import DataType
+from pyspark.sql.types import FloatType
 from pyspark.sql.types import IntegerType
 from pyspark.sql.types import StringType
 from pyspark.sql.types import StructField
@@ -274,7 +275,7 @@ class StructureMap_SourceSchema:
                 # A value to use if there is no existing value in the source object.
                 StructField("defaultValueDateTime", StringType(), True),
                 # A value to use if there is no existing value in the source object.
-                StructField("defaultValueDecimal", IntegerType(), True),
+                StructField("defaultValueDecimal", FloatType(), True),
                 # A value to use if there is no existing value in the source object.
                 StructField("defaultValueId", StringType(), True),
                 # A value to use if there is no existing value in the source object.
